Very promising alpha-beta benchmark

+++ ZobristHash.h
@@ -13,9 +13,9 @@ class Zob_Hash
 {
 public:
 	Zob_Hash();
-	void addEntry(const u64 Key, float  value, int depth);
-	bool getEntry(const u64 Key, float& value, int& depth) const;
-	bool getEntry(const u64 key, float& value) const;
+	void addEntry(const u64 Key, int  value, int depth);
+	bool getEntry(const u64 Key, int& value, int& depth) const;
+	bool getEntry(const u64 key, int& value) const;
 	bool hasEntry(const u64) const;
 	vector<vector<u64>> getRandomSet(); // Used once for init
 private:
